+++ src/config.py
@@ -1,0 +1,7 @@
+STEP_SHP_FOLDER = '/home/pcl/Data/GBA/db'
+STEP_CSV_FOLDER = '/home/pcl/Data/GBA/db/step'
+
+
+if __name__ == '__main__':
+    print(STEP_CSV_FOLDER)
+